+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes are documented in this file.
 ## *TBD* - *TBD*
 
 - Corrections and clarifications in the documentation
+- Corrections and clarifications in the maintenance cheat-sheet
 
 ## v20.1 - 2024-06-19